8.02.2021

+++ Business/Abstract/ICarService.cs
@@ -11,7 +11,7 @@ namespace Business.Abstract
         void Add(Car car);
         List<Car> GetAll();
         List<Car> GetAllByBrandId(int id);
-        List<Car> GetByDailyPrice(decimal min, decimal max);
+        List<Car> GetByDailyPrice(int min, int max);
         List<CarDetailDto> GetCarDetails(); 
     }
 }
